change: re-organize tests.base to make easier to maintain it
