+++ tests/base/__init__.py
@@ -5,18 +5,18 @@
 r"""Common test Utility functions, constants and global variables, etc.
 """
 from .common import TESTS_DIR, RES_DIR, NULL_CNTNR
-from .collector import TData, TDataCollector
-from .datasets import DATA_00
+from .collector import TDataCollector
+from .datatypes import TDataPaths, TData
 from .utils import (
-    resource_path, list_resources
+    load_data, each_data_from_dir
 )
 
 
 __all__ = [
     'TESTS_DIR', 'RES_DIR', 'NULL_CNTNR',
-    'TData', 'TDataCollector',
-    'DATA_00',
-    'resource_path', 'list_resources',
+    'TDataCollector',
+    'TDataPaths', 'TData',
+    'load_data', 'each_data_from_dir'
 ]
 
 # vim:sw=4:ts=4:et:
